+++ libft/ft_intlen.c
@@ -6,7 +6,7 @@
 /*   By: zshanabe <zshanabe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/26 19:55:56 by zshanabe          #+#    #+#             */
-/*   Updated: 2018/05/20 18:47:28 by zshanabe         ###   ########.fr       */
+/*   Updated: 2018/05/21 19:00:33 by zshanabe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int		ft_intlen(intmax_t n)
 {
 	int len;
-	int num;
+	intmax_t num;
 
 	num = n;
 	len = 1;
